# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Indonesian Stock Analysis API",
    "description": "Production-grade realtime and historical analysis API for Indonesian (IDX) equities.\n\nAll `/api/v1` data endpoints require authentication (`X-API-Key` header or `Authorization: Bearer <jwt>`) **and** an allow-listed browser origin. Responses use a uniform `{meta, data}` envelope; errors use a uniform `{error}` envelope.",
    "contact": {
      "name": "teguh-prasetyo.com",
      "url": "https://www.teguh-prasetyo.com/"
    },
    "license": {
      "name": "Proprietary"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://api.teguh-prasetyo.com",
      "description": "Production"
    },
    {
      "url": "http://localhost:8000",
      "description": "Local development"
    }
  ],
  "paths": {
...
    "/api/v1/auth/token": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Issue access & refresh tokens",
        "operationId": "issue_token_api_v1_auth_token_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Exchange a refresh token",
        "operationId": "refresh_token_api_v1_auth_refresh_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/technical/{symbol}": {
      "get": {
        "tags": [
          "analysis"
        ],
        "summary": "Full technical-indicator bundle",
        "operationId": "get_technical_api_v1_technical__symbol__get",
        "security": [
          {
            "APIKeyHeader": []
          },
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 12,
              "description": "IDX ticker symbol.",
              "examples": [
                "BBCA"
              ],
              "title": "Symbol"
            },
            "description": "IDX ticker symbol."
          },
          {
            "name": "interval",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/Interval",
              "description": "Sampling interval.",
              "default": "1d"
            },
            "description": "Sampling interval."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_TechnicalIndicators_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/signal/{symbol}": {
      "get": {
        "tags": [
          "analysis"
        ],
        "summary": "Composite trading signal with confidence and risk",
        "operationId": "get_signal_api_v1_signal__symbol__get",
        "security": [
          {
            "APIKeyHeader": []
          },
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 12,
              "description": "IDX ticker symbol.",
              "examples": [
                "BBCA"
              ],
              "title": "Symbol"
            },
            "description": "IDX ticker symbol."
          },
          {
            "name": "interval",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/Interval",
              "description": "Sampling interval.",
              "default": "1d"
            },
            "description": "Sampling interval."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_TradingSignal_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/market/summary": {
      "get": {
        "tags": [
          "market"
        ],
        "summary": "Aggregate snapshot of the tracked IDX universe",
        "operationId": "market_summary_api_v1_market_summary_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_MarketSummary_"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          },
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/top-gainers": {
      "get": {
        "tags": [
          "market"
        ],
        "summary": "Top gainers by percentage change",
        "operationId": "top_gainers_api_v1_top_gainers_get",
        "security": [
          {
            "APIKeyHeader": []
          },
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 45,
              "minimum": 1,
              "default": 10,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_list_MarketMover__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/top-losers": {
      "get": {
        "tags": [
          "market"
        ],
        "summary": "Top losers by percentage change",
        "operationId": "top_losers_api_v1_top_losers_get",
        "security": [
          {
            "APIKeyHeader": []
          },
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 45,
              "minimum": 1,
              "default": 10,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_list_MarketMover__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/unusual-volume": {
      "get": {
        "tags": [
          "market"
        ],
        "summary": "Symbols trading on unusually high relative volume",
        "operationId": "unusual_volume_api_v1_unusual_volume_get",
        "security": [
          {
            "APIKeyHeader": []
          },
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 45,
              "minimum": 1,
              "default": 10,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_list_MarketMover__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Envelope_MarketSummary_": {
        "properties": {
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          },
          "data": {
            "$ref": "#/components/schemas/MarketSummary"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "meta",
          "data"
        ],
        "title": "Envelope[MarketSummary]"
      },
...
      "Envelope_TechnicalIndicators_": {
        "properties": {
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          },
          "data": {
            "$ref": "#/components/schemas/TechnicalIndicators"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "meta",
          "data"
        ],
        "title": "Envelope[TechnicalIndicators]"
      },
      "Envelope_TradingSignal_": {
        "properties": {
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          },
          "data": {
            "$ref": "#/components/schemas/TradingSignal"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "meta",
          "data"
        ],
        "title": "Envelope[TradingSignal]"
      },
      "Envelope_list_MarketMover__": {
        "properties": {
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/MarketMover"
            },
            "type": "array",
            "title": "Data"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "meta",
          "data"
        ],
        "title": "Envelope[list[MarketMover]]"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Interval": {
        "type": "string",
        "enum": [
          "1m",
          "5m",
          "15m",
          "30m",
          "1h",
          "1d",
          "1wk",
          "1mo"
        ],
        "title": "Interval",
        "description": "Supported sampling intervals (mapped to provider-native values)."
      },
      "MACDValue": {
        "properties": {
          "macd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Macd"
          },
          "signal": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Signal"
          },
          "histogram": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Histogram"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "MACDValue"
      },
      "MarketMover": {
        "properties": {
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "price": {
            "type": "number",
            "title": "Price"
          },
          "change": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Change"
          },
          "change_percent": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Change Percent"
          },
          "volume": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Volume"
          },
          "relative_volume": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Relative Volume"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "symbol",
          "price"
        ],
        "title": "MarketMover",
        "description": "A single row in a market screener (gainers/losers/unusual volume)."
      },
      "MarketSummary": {
        "properties": {
          "universe_size": {
            "type": "integer",
            "title": "Universe Size"
          },
          "evaluated": {
            "type": "integer",
            "title": "Evaluated"
          },
          "advancers": {
            "type": "integer",
            "title": "Advancers"
          },
          "decliners": {
            "type": "integer",
            "title": "Decliners"
          },
          "unchanged": {
            "type": "integer",
            "title": "Unchanged"
          },
          "average_change_percent": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Average Change Percent"
          },
          "is_market_open": {
            "type": "boolean",
            "title": "Is Market Open"
          },
          "top_gainers": {
            "items": {
              "$ref": "#/components/schemas/MarketMover"
            },
            "type": "array",
            "title": "Top Gainers"
          },
          "top_losers": {
            "items": {
              "$ref": "#/components/schemas/MarketMover"
            },
            "type": "array",
            "title": "Top Losers"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "universe_size",
          "evaluated",
          "advancers",
          "decliners",
          "unchanged",
          "is_market_open",
          "top_gainers",
          "top_losers"
        ],
        "title": "MarketSummary",
        "description": "Aggregate snapshot of the tracked IDX universe."
      },
...
      "RefreshRequest": {
        "properties": {
          "refresh_token": {
            "type": "string",
            "maxLength": 4096,
            "minLength": 16,
            "title": "Refresh Token"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "refresh_token"
        ],
        "title": "RefreshRequest"
      },
      "ResponseMeta": {
        "properties": {
          "symbol": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Symbol",
            "description": "Canonical IDX symbol, e.g. BBCA.JK."
          },
          "exchange": {
            "type": "string",
            "title": "Exchange",
            "description": "Source exchange identifier.",
            "default": "IDX"
          },
          "currency": {
            "type": "string",
            "title": "Currency",
            "description": "Currency of all monetary values.",
            "default": "IDR"
          },
          "timezone": {
            "type": "string",
            "title": "Timezone",
            "description": "Time zone of all timestamps.",
            "default": "Asia/Jakarta"
          },
          "source": {
            "type": "string",
            "title": "Source",
            "description": "Upstream data provider that served the data."
          },
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At",
            "description": "UTC instant the response was produced."
          },
          "cached": {
            "type": "boolean",
            "title": "Cached",
            "description": "Whether the payload was served from cache.",
            "default": false
          },
          "disclaimer": {
            "type": "string",
            "title": "Disclaimer",
            "default": "Data is provided for informational purposes only and may be delayed. It is not investment advice."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "source",
          "generated_at"
        ],
        "title": "ResponseMeta",
        "description": "Provenance metadata attached to every successful data response."
      },
...
      "SignalType": {
        "type": "string",
        "enum": [
          "strong_buy",
          "buy",
          "hold",
          "sell",
          "strong_sell"
        ],
        "title": "SignalType",
        "description": "Discrete trading signal classification."
      },
...
      "TechnicalIndicators": {
        "properties": {
          "last_price": {
            "type": "number",
            "title": "Last Price"
          },
          "moving_averages": {
            "$ref": "#/components/schemas/MovingAverages"
          },
          "rsi_14": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rsi 14"
          },
          "macd": {
            "$ref": "#/components/schemas/MACDValue"
          },
          "bollinger_bands": {
            "$ref": "#/components/schemas/BollingerBands"
          },
          "atr_14": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Atr 14"
          },
          "annualized_volatility": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Annualized Volatility"
          },
          "support_resistance": {
            "$ref": "#/components/schemas/SupportResistance"
          },
          "trend": {
            "$ref": "#/components/schemas/TrendInfo"
          },
          "volume": {
            "$ref": "#/components/schemas/VolumeAnalysis"
          },
          "sample_size": {
            "type": "integer",
            "title": "Sample Size",
            "description": "Number of bars used in the computation."
          },
          "interval": {
            "$ref": "#/components/schemas/Interval"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "last_price",
          "moving_averages",
          "macd",
          "bollinger_bands",
          "support_resistance",
          "trend",
          "volume",
          "sample_size",
          "interval"
        ],
        "title": "TechnicalIndicators",
        "description": "The full technical-indicator bundle for a symbol."
      },
      "TokenRequest": {
        "properties": {
          "scopes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 16,
            "title": "Scopes",
            "description": "Subset of grantable scopes to embed in the access token."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "TokenRequest",
        "description": "Request body to mint tokens. Authentication itself is via the API key\nheader; this body only narrows the requested scopes."
      },
      "TokenResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "refresh_token": {
            "type": "string",
            "title": "Refresh Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          },
          "expires_in": {
            "type": "integer",
            "title": "Expires In",
            "description": "Access-token lifetime in seconds."
          },
          "scopes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Scopes"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "access_token",
          "refresh_token",
          "expires_in",
          "scopes"
        ],
        "title": "TokenResponse"
      },
      "TradingSignal": {
        "properties": {
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "signal": {
            "$ref": "#/components/schemas/SignalType"
          },
          "confidence": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence"
          },
          "risk_level": {
            "$ref": "#/components/schemas/RiskLevel"
          },
          "score": {
            "type": "number",
            "title": "Score",
            "description": "Net weighted score in the range [-1, 1]."
          },
          "trend": {
            "$ref": "#/components/schemas/TrendDirection"
          },
          "components": {
            "items": {
              "$ref": "#/components/schemas/SignalComponent"
            },
            "type": "array",
            "title": "Components"
          },
          "interval": {
            "$ref": "#/components/schemas/Interval"
          },
          "evaluated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Evaluated At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "symbol",
          "signal",
          "confidence",
          "risk_level",
          "score",
          "trend",
          "components",
          "interval",
          "evaluated_at"
        ],
        "title": "TradingSignal",
        "description": "Composite trading signal with confidence and risk classification."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      },
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}